feat: add optional LVGL UI layer and architectural refinements - Add oc::ui module with LVGLBridge, IElement/IWidget/IComponent/IView   interfaces, and LVGLAdapter for scoped bindings - Move EncoderLogic to oc/core/input/ (core business logic) - Extract shared EncoderDef to drivers/common/ (reduce duplication) - Convert Teensy EncoderController to template for API parity with Arduino - Add std::atomic<bool> for ISR-safe pending flag in EncoderLogic - Wire MIDI IN callbacks to EventBus in OpenControlApp - Implement lazy context initialization in ContextManager - Remove unused setFlushCallback from IDisplayDriver - Add teensy41-dev environment with LVGL for testing - Move lv_conf.h to config/dev/ (consumer provides their own) Completes Phases 7, 7.1, and 7.2 of migration plan.

+++ src/oc/hal/Types.hpp
@@ -29,12 +29,14 @@ enum class ButtonEvent : uint8_t {
 
 /**
  * @brief Rectangle for display regions
+ *
+ * Uses int32_t for compatibility with LVGL 9 coordinates.
  */
 struct Rect {
-    int16_t x1;
-    int16_t y1;
-    int16_t x2;
-    int16_t y2;
+    int32_t x1;
+    int32_t y1;
+    int32_t x2;
+    int32_t y2;
 };
 
 /**
